backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic>=2.7.0
pydantic-settings>=2.3.0
sqlalchemy>=2.0.0
aiosqlite==0.20.0
ofxtools==0.9.4
python-multipart==0.0.9
python-dotenv==1.0.1
alembic==1.13.2
httpx==0.27.0
# v2 additions
python-dateutil>=2.9.0       # relativedelta for installment date arithmetic
cryptography>=42.0.0         # AES-256-GCM field-level encryption
